🚀 3단계 - 수강신청(DB 적용)

+++ src/main/java/nextstep/session/domain/SessionRepository.java
@@ -2,7 +2,7 @@ package nextstep.session.domain;
 
 public interface SessionRepository {
 
-    int save(Session session);
+    Long save(Session session);
 
     Session findById(Long id);
